requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
requests==2.31.0
openai>=1.55.3
httpx>=0.27.2
python-dotenv==1.0.0
boto3==1.34.0
pdfplumber==0.10.3
openpyxl==3.1.2
python-docx==1.1.0
Pillow==10.1.0
pandas==2.1.4
aiofiles==23.2.1
psutil==5.9.8

